`Fix `<IsPackable>` value and add license file reference`
Corrected the value of the `<IsPackable>` property from `trued` to `true` to ensure the project can be packed. Added a new `<PackageLicenseFile>` property to the `<PropertyGroup>` to specify the license file for the package. Also, included a new `<ItemGroup>` with a reference to the license file (`LICENSE.txt`) set to be packed with the project and placed in the root directory of the package.